refactor(settings): improve folder structure of components
