+++ src/components/Settings/LegalList.tsx
