+++ src/components/Settings/__tests__/LegalList.test.tsx
